MIP-314 Updated diagram, refactored model file locations
