+++ src/main/groovy/com/avioconsulting/mule/linter/rule/configuration/GlobalConfigNoFlowsRule.groovy
@@ -1,9 +1,9 @@
 package com.avioconsulting.mule.linter.rule.configuration
 
 import com.avioconsulting.mule.linter.model.Application
-import com.avioconsulting.mule.linter.model.MuleComponent
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.configuration.MuleComponent
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 
 class GlobalConfigNoFlowsRule extends Rule {
