+++ src/main/groovy/com/avioconsulting/mule/linter/rule/git/GitIgnoreRule.groovy
@@ -2,8 +2,8 @@ package com.avioconsulting.mule.linter.rule.git
 
 import com.avioconsulting.mule.linter.model.Application
 import com.avioconsulting.mule.linter.model.GitIgnoreFile
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 
 class GitIgnoreRule extends Rule {
