+++ src/main/groovy/com/avioconsulting/mule/linter/rule/muleartifact/MuleArtifactHasSecurePropertiesRule.groovy
@@ -1,8 +1,8 @@
 package com.avioconsulting.mule.linter.rule.muleartifact
 
 import com.avioconsulting.mule.linter.model.Application
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 import org.apache.groovy.json.internal.JsonArray
 
 class MuleArtifactHasSecurePropertiesRule extends Rule {
