+++ src/main/groovy/com/avioconsulting/mule/linter/rule/pom/MuleRuntimeVersionRule.groovy
@@ -1,7 +1,7 @@
 package com.avioconsulting.mule.linter.rule.pom
 
 import com.avioconsulting.mule.linter.model.Application
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 
 class MuleRuntimeVersionRule extends PomPropertyValueRule {
