+++ src/main/groovy/com/avioconsulting/mule/linter/rule/pom/MunitMavenPluginAttributesRule.groovy
@@ -1,8 +1,8 @@
 package com.avioconsulting.mule.linter.rule.pom
 
 import com.avioconsulting.mule.linter.model.Application
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 import com.avioconsulting.mule.linter.model.pom.MunitMavenPlugin
 import com.avioconsulting.mule.linter.model.pom.PomElement
 import com.avioconsulting.mule.linter.model.pom.PomFile
