+++ src/main/groovy/com/avioconsulting/mule/linter/rule/pom/PomPropertyValueRule.groovy
@@ -2,8 +2,8 @@ package com.avioconsulting.mule.linter.rule.pom
 
 import com.avioconsulting.mule.linter.model.Application
 import com.avioconsulting.mule.linter.model.pom.PomElement
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 
 class PomPropertyValueRule extends Rule {
