+++ src/main/groovy/com/avioconsulting/mule/linter/rule/property/EncryptedPasswordRule.groovy
@@ -2,8 +2,8 @@ package com.avioconsulting.mule.linter.rule.property
 
 import com.avioconsulting.mule.linter.model.Application
 import com.avioconsulting.mule.linter.model.PropertyFile
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 
 class EncryptedPasswordRule extends Rule {
