+++ src/test/groovy/com/avioconsulting/mule/linter/cicd/JenkinsFileExistsRuleTest.groovy
@@ -3,8 +3,8 @@ package com.avioconsulting.mule.linter.cicd
 import com.avioconsulting.mule.linter.TestApplication
 import com.avioconsulting.mule.linter.model.Application
 import com.avioconsulting.mule.linter.model.JenkinsFile
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 import com.avioconsulting.mule.linter.rule.cicd.JenkinsFileExistsRule
 import spock.lang.Specification
