+++ src/test/groovy/com/avioconsulting/mule/linter/rule/configuration/ConfigFileNamingRuleTest.groovy
@@ -3,8 +3,8 @@ package com.avioconsulting.mule.linter.rule.configuration
 import com.avioconsulting.mule.linter.TestApplication
 import com.avioconsulting.mule.linter.model.Application
 import com.avioconsulting.mule.linter.model.CaseNaming
-import com.avioconsulting.mule.linter.model.Rule
-import com.avioconsulting.mule.linter.model.RuleViolation
+import com.avioconsulting.mule.linter.model.rule.Rule
+import com.avioconsulting.mule.linter.model.rule.RuleViolation
 import spock.lang.Specification
 
 class ConfigFileNamingRuleTest extends Specification {
